[BLEU] Support variable number of references
When evaluating with multi-bleu.perl with multiple reference files,
having one or more empty references for a given translation could cause
the brevity penalty to be miscalculated. The closest length could indeed
be the one of the empty reference.

Example:
+-------+--------------------+--------------------------+------+
| trans | ref1               | ref2                     | ref3 |
+-------+--------------------+--------------------------+------+
| I do  | this is a long ref | this is another long ref |      |
+-------+--------------------+--------------------------+------+

Here ref3 (empty / lenght=0) should not be considered for finding
the reference with closest length for the brevity penalty.